# backend/api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lead Generation Platform API",
    "description": "API for enriching domains with decision-maker contacts. Uses internal Contacts DB first, falls back to Blitz API, and automatically syncs results back to the internal database.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@eagleinfoservice.com"
    }
  },
  "servers": [
    {
      "url": "https://listbuilding.eagleinfoservice.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8765",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login to get JWT token",
        "description": "Authenticate with email and password to receive a JWT token for API access.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "your_password"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful - returns JWT token",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user_id": "550e8400-e29b-41d4-a716-446655440000",
                  "email": "user@example.com",
                  "is_admin": false
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid email or password."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh JWT token",
        "description": "Get a new JWT token with extended expiration. Requires valid existing token.",
        "operationId": "refresh_token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user_id": "550e8400-e29b-41d4-a716-446655440000",
                  "email": "user@example.com",
                  "is_admin": false
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user info",
        "description": "Returns the profile of the currently authenticated user.",
        "operationId": "get_current_user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "example": {
                  "user_id": "550e8400-e29b-41d4-a716-446655440000",
                  "email": "user@example.com",
                  "is_admin": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      }
    }
  }
}